combining chainport transaction logic
